Back porting clean up of FlowFrameworkTests.kt made in ENT
